fix: remove ONNX file loading functionality and simplify to built-in models only
- Delete src/onnx_models.rs module completely
- Remove ONNX file path loading from src/model.rs
- Update src/main.rs to use built-in model names instead of file paths
- Replace validate_model_path with validate_built_in_model_name
- Fix all type mismatches (String vs PathBuf)
- Remove unused imports and clean up code
- Simplify test cases to use dummy models only
- Keep burn-import for generating embedded models but remove runtime ONNX loading

Now the system only supports built-in models compiled into the binary,
no external ONNX file loading at runtime.
